Added use of DataManager
Now every onSensorChange(), lastTime dataword is set to thisTime
dataWord, and a new DataWord is initiated for thisTime.
Then thisTime is passed to DataWord.addMAINDATA() so that every sensor
update’s data is stored.
TODO: At the moment after a short period of time the RAM will fill with
the update information, I will change it so that information is written
onto a file.